GYRE-1221 fix2

+++ app/Http/GraphQL/Mutations/RemoveSocialConnect.php
@@ -43,6 +43,9 @@ class RemoveSocialConnect extends Mutation
             ->where('user_id', '=', $this->getGuard()->user()->id)
             ->first()
         ;
+        if (empty($social)) {
+            return null;
+        }
         $social->forceDelete();
 
         $socialRepository = new SocialRepository();
